interceptor vfx files, actions, abillity file starter pack
Added the vfx and vfx textures for GK interceptors (6 vfx total, a 15s timer, and 5 segments each lasting 3 seconds).

Action files for all 6 versions are present, and an ability file "interceptor timer" which was left blank at first has all 6 actions in there in the correct format to be used later/saved.